warning fix
git-svn-id: http://svn.miranda-ng.org/main/trunk@8628 1316c22d-e87f-b044-9b9b-93d7a3e3ba9c

+++ protocols/FacebookRM/src/utils.cpp
@@ -456,7 +456,7 @@ std::string utils::text::rand_string(int len, const char *chars)
 
 	srand(::time(NULL));
 	for (int i = 0; i < len; ++i) {
-		out << chars[utils::number::random(0, strlen(chars))];
+		out << chars[utils::number::random(0, (int)strlen(chars))];
 	}
 
 	return out.str();
